Create FloydWarshall.java
The Floyd-Warshall algorithm is used to find the shortest paths between all pairs of vertices in a weighted graph. Here it is  a Java implementation of the Floyd-Warshall algorithm!
This Java code demonstrates the Floyd-Warshall algorithm for finding the shortest paths between all pairs of vertices in a weighted graph. Make sure to adjust the graph array and the number of vertices (V) to match your specific graph.